+++ lib/src/commands/version_mixin.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:http/http.dart';
 import 'package:path/path.dart';
@@ -11,10 +12,21 @@ import 'package:pub_semver/pub_semver.dart';
 import 'package:upcode_ci/src/commands/command.dart';
 
 mixin VersionMixin on UpcodeCommand {
-  String get versionType => split(flutterDir).last;
+  String get versionType {
+    return split(normalize(File(flutterDir).absolute.path)).last;
+  }
 
   String get databaseUrl {
     final String databaseKey = join(privateDir, 'firebase_database.key').readAsStringSync();
+
+    final Map<String, dynamic> data =
+        Map<String, dynamic>.from(jsonDecode(join(androidAppDir, 'google-services.json').readAsStringSync()));
+
+    final String url = data['project_info'] == null ? null : data['project_info']['firebase_url'];
+    if (url != null) {
+      return '${url}/$versionType/.json?auth=$databaseKey';
+    }
+
     return 'https://$projectId.firebaseio.com/$versionType/.json?auth=$databaseKey';
   }
 
@@ -26,7 +38,7 @@ mixin VersionMixin on UpcodeCommand {
   }
 
   Future<void> setVersion(Version version) async {
-    await patch(
+    final result = await patch(
       databaseUrl,
       body: jsonEncode(
         <String, dynamic>{
@@ -35,5 +47,8 @@ mixin VersionMixin on UpcodeCommand {
         },
       ),
     );
+
+    print(databaseUrl);
+    print(result.body);
   }
 }